further refactoring of code base. THis branch is becoming a refactoring branch
