+++ NFACT/NFACT_decomp/__main__.py
@@ -3,6 +3,8 @@ import shutil
 
 from NFACT.NFACT_base.utils import Timer, colours
 from NFACT.NFACT_base.signithandler import Signit_handler
+from NFACT.NFACT_base.imagehandling import save_images
+
 from NFACT.NFACT_decomp.setup.args import nfact_args
 from NFACT.NFACT_decomp.setup.file_setup import (
     create_folder_set_up,
@@ -22,7 +24,7 @@ from NFACT.NFACT_decomp.decomposition.matrix_handling import (
     load_previous_matrix,
     save_avg_matrix,
 )
-from NFACT.NFACT_decomp.pipes.image_handling import save_images
+
 from NFACT.NFACT_decomp.pipes.image_handling import winner_takes_all
 from NFACT.NFACT_decomp.setup.arg_check import (
     check_complusory_arguments,
